Remove all zoom-related code; completely breaks printing.
When the sheets are scaled, including the auto-scale, it breaks the scaling for
print media. Print CSS does not properly transform back, JavaScript cannot
reliably intercept print calls in order to try and correct scaling before the
print dialog presents itself. While the zooming controls and auto-scaling were
great usability features, reliable printing is a higher priority. This is
something that can possibly be re-implemented in the future.