Removing print button. Print support is more broken than expected.
After extensive testing, it is apparent that the printing support is extremely
broken. There are too many variables with browser size, margins, scaling, etc,
to be able to expect the user to be able to reliably print with good results.
The solution to this would be to generate our own pdf using a library like
jsPDF. The problem is the libraries it uses to generate HTML elements,
html2canvas or rasterizeHTML, are very limited in SVG and custom CSS support,
both of which are relied on heavily for this project. I only see this being able
to be accomplished by generating images to include in a PDF by making our own
canvas elements. I am going to have to do quite a bit of research and testing
before I will be able to accomplish this, so it may be some time before
printing/pdf export is available.

Good news is, since we will not rely on browser print support, we can re-include
element scaling for zoom support.